Refactor K6 tests and utilities
- Remove redundant comments and annotations from thresholds and `shelley-era-endpoints`.
- Relocate `buildUrl` to `helpers.ts` and update imports across scenarios.
- Improve readability of helper functions with consistent formatting.
- Simplify and consolidate Shelley-specific endpoint configurations.
