+++ tests/k6/src/scenarios/accounts.ts
@@ -1,7 +1,8 @@
-import { buildUrl, ENDPOINTS } from '../config/endpoints';
+import { ENDPOINTS } from '../config/endpoints';
 import { getRandomItem } from '../config/test-data';
 import { TEST_DATA } from '../config/shelley-test-data';
 import { apiClient, MetricType } from '../utils/api-client';
+import { buildUrl } from '../utils/helpers';
 
 export function testGetAccount(): void {
   const stakeAddress = getRandomItem(TEST_DATA.stakeAddresses);
